Set type as attr, use name as reference (#71)
* config: change type to attribute

Also make compressor required. We'll get a panic if no compressor is not
set, but in reality it should always be set.

* graph: support type attribute and name as ref

Update decoder to support type attribute, based on changes in config.

In addition, references to other resources now only use the name.
Previously, the <type>.<name> combination was used to reference an
object. Now only the name is used. An error is returned if two resources
have the same name.

In practice, this means:

  resource "a" "foo" {
    input = "hello"
  }
  resource "b" "bar" {
    input = "${a.foo.input}"
  }

becomes

  resource "foo" {
    type  = "a"
    input = "hello"
  }
  resource "bar" {
    type  = "b"
    input = "${foo.input}"
  }

* provider/aws: rename conflicting type attribute