graph: record deps before resolving dependencies (#75)
This fixes an issue where expression that could be statically resolved
would not record a dependency relationship, resulting in in child
resources to be deleted concurrently with the parent resource.

Previously, the dependencies were set based on what was set in the
graph. With this commit, the dependencies are set based on the user
specified config instead.

With this change, deletion will happen slower as fewer tasks can run
concurrently. However, it guarantees that resources are always deleted
in the correct order.