resource: add input/output to resource for config
For now, the value does nothing. This is the first step in transitioning
the resource config to be a cty.Value, rather than an instance of a
resource.Definition.

Several tests use cmp to compare values that contain contain resources
and must be updated as the cty.Value cannot be compared as is (due to
unexported fields).