resource/graph: implement new graph package
The new graph package, intended to replace graph, is a rewrite of graph
and related packages with the primary goal of making the graph
serializable.

Another goal of the graph was to make it easier to construct and compare
in tests. This means the snapshot package is no longer required.