resource/graph: use resource json encoder (#91)
JSON encoding for resources was implemented twice. This commit uses the
same encoder for encoding resources in the graph as what's used when
encoding individual resources.

Previously, the type was inferred from json, which could result in the types not
matching. Now, when encoding and decoding, the registry is required to ensure
that types match.